+++ cypress/fixtures/outline.feature.ts
@@ -1,4 +1,4 @@
-import { feature, scenarioOutline, scenario, Given, given, When, when, Then, then } from '@hypercubed/cypress-gherkin';
+import { feature, scenarioOutline, outline, given, when, then, examples, Given, When, Then } from '@hypercubed/cypress-gherkin';
 
 Given('there are {int} cucumbers', (a: number) => {
 
@@ -14,32 +14,22 @@ Then('I should have {int} cucumbers', (a: number) => {
 
 feature('Examples Tables in Gherkin text', () => {
   scenarioOutline('eating cucumbers', () => {
-    scenarioOutline('These are passing', () => {
-      scenario('example', () => {
-        given('there are 12 cucumbers');
-        when('I eat 5 cucumbers');
-        then('I should have 7 cucumbers');
-      });
-      
-      scenario('example', () => {
-        given('there are 20 cucumbers');
-        when('I eat 5 cucumbers');
-        then('I should have 15 cucumbers');
-      });
+    outline(() => {
+      given('there are <start> cucumbers');
+      when('I eat <eat> cucumbers');
+      then('I should have <left> cucumbers');
     });
-    
-    scenarioOutline('These are also passing', () => {
-      scenario('example', () => {
-        given('there are 22 cucumbers');
-        when('I eat 5 cucumbers');
-        then('I should have 17 cucumbers');
-      });
-      
-      scenario('example', () => {
-        given('there are 10 cucumbers');
-        when('I eat 5 cucumbers');
-        then('I should have 5 cucumbers');
-      });
-    });
+
+    examples('These are passing', [
+      ['start', 'eat', 'left'],
+      [12, 5, 7],
+      [20, 5, 15]
+    ]);
+
+    examples('These are also passing', [
+      ['start', 'eat', 'left'],
+      [22, 5, 17],
+      [10, 5, 5]
+    ]);
   });
 });